+++ libs/core/core/components/apps/base-app/base-app-routes.tsx
@@ -35,7 +35,7 @@ import ConfiguredDevicesDiscovery from "Core/discovery-device/components/configu
 import DevicesInitialization from "Core/device-initialization/components/devices-initialization.component"
 import AvailableDeviceListContainer from "Core/discovery-device/components/available-device-list.container"
 import DeviceConnecting from "Core/discovery-device/components/device-connecting.component"
-import { ApiDeviceModals, GenericView } from "generic-view/feature"
+import { GenericView } from "generic-view/feature"
 import { APIConnectionDemo } from "generic-view/ui"
 
 // AUTO DISABLED - fix me if you like :)
@@ -130,6 +130,5 @@ export default () => (
       </Route>
       <Redirect to={URL_OVERVIEW.root} />
     </Switch>
-    <ApiDeviceModals />
   </>
 )
