+++ libs/generic-view/feature/src/lib/api-device-modals.tsx
@@ -5,20 +5,8 @@
 
 import React, { FunctionComponent } from "react"
 import { GenericThemeProvider } from "generic-view/theme"
-import {
-  BackupError,
-  BackupRestoreError,
-  ModalBase,
-  ModalCenteredContent,
-} from "generic-view/ui"
-import { useDispatch, useSelector } from "react-redux"
-import {
-  cleanBackupProcess,
-  cleanRestoreProcess,
-  selectActiveDevice,
-  selectBackupProcessStatus,
-  selectBackupRestoreStatus,
-} from "generic-view/store"
+import BackupErrorModal from "./modals/backup-error-modal"
+import RestoreErrorModal from "./modals/restore-error-modal"
 
 export const ApiDeviceModals: FunctionComponent = () => {
   return (
@@ -26,43 +14,5 @@ export const ApiDeviceModals: FunctionComponent = () => {
       <BackupErrorModal />
       <RestoreErrorModal />
     </GenericThemeProvider>
-  )
-}
-
-const BackupErrorModal: FunctionComponent = () => {
-  const dispatch = useDispatch()
-  const activeDevice = useSelector(selectActiveDevice)
-  const backupStatus = useSelector(selectBackupProcessStatus)
-  const opened = backupStatus === "FAILED" && !activeDevice
-
-  const onClose = () => {
-    dispatch(cleanBackupProcess())
-  }
-  return (
-    <ModalBase opened={opened} variant={"small"}>
-      <ModalCenteredContent>
-        <BackupError closeAction={{ type: "custom", callback: onClose }} />
-      </ModalCenteredContent>
-    </ModalBase>
-  )
-}
-
-const RestoreErrorModal: FunctionComponent = () => {
-  const dispatch = useDispatch()
-  const activeDevice = useSelector(selectActiveDevice)
-  const restoreStatus = useSelector(selectBackupRestoreStatus)
-  const opened = restoreStatus === "FAILED" && !activeDevice
-
-  const onClose = () => {
-    dispatch(cleanRestoreProcess())
-  }
-  return (
-    <ModalBase opened={opened} variant={"small"}>
-      <ModalCenteredContent>
-        <BackupRestoreError
-          closeAction={{ type: "custom", callback: onClose }}
-        />
-      </ModalCenteredContent>
-    </ModalBase>
   )
 }
